+++ thunderstore/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+### 0.2.0
+* A border is now drawn around the bunny hood in the pause menu when equipped
+* Postman timer SFX is now correctly handled
+* Postman timer no longer displays in situations where it shouldn't
+
 ### 0.1.0
 * Initial release